moved common saml functionality into core
