+++ cas-server-support-saml-core/build.gradle
@@ -1,0 +1,20 @@
+description = 'Apereo CAS SAML Core/Common Functionality'
+dependencies {
+    compile libraries.opensaml
+    compile libraries.shibidp
+    compile libraries.bouncycastle
+    compile project(':cas-server-core-web')
+    compile project(':cas-server-core-services')
+    compile project(':cas-server-core-authentication')
+    compile project(':cas-server-webapp-validation')
+    compile libraries.jdom
+    runtime libraries.hibernate
+    runtime libraries.xmlapis
+    testCompile project(':cas-server-core-util')
+    testCompile project(':cas-server-core')
+    testCompile project(':cas-server-core-logout')
+    testCompile libraries.log4j
+    testCompile project(path: ":cas-server-core-authentication", configuration: "tests")
+    testCompile project(path: ":cas-server-core-services", configuration: "tests")
+    testCompile libraries.metrics
+}
