+++ cas-server-support-saml-core/src/main/java/org/jasig/cas/support/saml/OpenSamlConfigBean.java
@@ -18,7 +18,7 @@ import javax.annotation.PostConstruct;
 import javax.validation.constraints.NotNull;
 
 /**
- * Load the opensaml config context.
+ * Load the OpenSAML config context.
  * @author Misagh Moayyed
  * @since 4.1
  */
