+++ cas-server-support-saml/src/main/java/org/jasig/cas/config/SamlConfiguration.java
@@ -1,25 +1,14 @@
 package org.jasig.cas.config;
 
-import com.sun.org.apache.xerces.internal.util.SecurityManager;
-import net.shibboleth.utilities.java.support.xml.BasicParserPool;
 import org.jasig.cas.authentication.support.CasAttributeEncoder;
 import org.jasig.cas.services.ServicesManager;
-import org.jasig.cas.support.saml.OpenSamlConfigBean;
 import org.jasig.cas.support.saml.web.view.Saml10FailureResponseView;
 import org.jasig.cas.support.saml.web.view.Saml10SuccessResponseView;
-import org.opensaml.core.xml.XMLObjectBuilderFactory;
-import org.opensaml.core.xml.config.XMLObjectProviderRegistrySupport;
-import org.opensaml.core.xml.io.MarshallerFactory;
-import org.opensaml.core.xml.io.UnmarshallerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.DependsOn;
-
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * This is {@link SamlConfiguration} that creates the necessary opensaml context and beans.
@@ -29,11 +18,6 @@ import java.util.Map;
  */
 @Configuration("samlConfiguration")
 public class SamlConfiguration {
-
-    /**
-     * The constant POOL_SIZE.
-     */
-    private static final int POOL_SIZE = 100;
 
     /**
      * The Issuer.
@@ -88,83 +72,5 @@ public class SamlConfiguration {
         view.setCasAttributeEncoder(this.casAttributeEncoder);
         return view;
     }
-
-    /**
-     * Open saml config bean open saml config bean.
-     *
-     * @return the open saml config bean
-     */
-    @Bean(name="shibboleth.OpenSAMLConfig")
-    @DependsOn("shibboleth.ParserPool")
-    public OpenSamlConfigBean openSamlConfigBean() {
-        return new OpenSamlConfigBean();
-    }
-
-    /**
-     * Parser pool basic parser pool.
-     *
-     * @return the basic parser pool
-     */
-    @Bean(name="shibboleth.ParserPool", initMethod = "initialize")
-    public BasicParserPool parserPool() {
-        final BasicParserPool pool = new BasicParserPool();
-        pool.setMaxPoolSize(POOL_SIZE);
-        pool.setCoalescing(true);
-        pool.setIgnoreComments(true);
-        pool.setXincludeAware(false);
-        pool.setExpandEntityReferences(false);
-        pool.setIgnoreComments(true);
-        pool.setNamespaceAware(true);
-        
-        final Map<String, Object> attributes = new HashMap<>();
-        attributes.put("http://apache.org/xml/properties/security-manager", 
-                new SecurityManager());
-        pool.setBuilderAttributes(attributes);
-        
-        final Map<String, Boolean> features = new HashMap<>();
-        features.put("http://apache.org/xml/features/disallow-doctype-decl", Boolean.TRUE);
-        features.put("http://apache.org/xml/features/validation/schema/normalized-value", 
-                Boolean.FALSE);
-        features.put("http://javax.xml.XMLConstants/feature/secure-processing", 
-                Boolean.TRUE);
-        features.put("http://xml.org/sax/features/external-general-entities",
-                Boolean.FALSE);
-        features.put("http://xml.org/sax/features/external-parameter-entities",
-                Boolean.FALSE);
-        pool.setBuilderFeatures(features);
-        return pool;
-    }
-
-    /**
-     * Builder factory xml object builder factory.
-     *
-     * @return the xml object builder factory
-     */
-    @Bean(name="shibboleth.BuilderFactory")
-    @DependsOn("shibboleth.OpenSAMLConfig")
-    public XMLObjectBuilderFactory builderFactory() {
-        return XMLObjectProviderRegistrySupport.getBuilderFactory();
-    }
-
-    /**
-     * Marshaller factory marshaller factory.
-     *
-     * @return the marshaller factory
-     */
-    @Bean(name="shibboleth.MarshallerFactory")
-    @DependsOn("shibboleth.OpenSAMLConfig")
-    public MarshallerFactory marshallerFactory() {
-        return XMLObjectProviderRegistrySupport.getMarshallerFactory();
-    }
-
-    /**
-     * Unmarshaller factory unmarshaller factory.
-     *
-     * @return the unmarshaller factory
-     */
-    @Bean(name="shibboleth.MarshallerFactory")
-    @DependsOn("shibboleth.OpenSAMLConfig")
-    public UnmarshallerFactory unmarshallerFactory() {
-        return XMLObjectProviderRegistrySupport.getUnmarshallerFactory();
-    }
+    
 }
