+++ cas-server-support-shibboleth-attributes/src/main/java/org/jasig/cas/config/ShibbolethAttributeResolverConfiguration.java
@@ -60,23 +60,4 @@ public class ShibbolethAttributeResolverConfiguration {
                 null
         );
     }
-
-    @Bean(name = "shibboleth.VelocityEngine")
-    VelocityEngineFactoryBean velocityEngineFactoryBean() {
-        final VelocityEngineFactoryBean velocityEngineFactoryBean = new VelocityEngineFactoryBean();
-
-        final Properties properties = new Properties();
-        properties.put("input.encoding", "UTF-8");
-        properties.put("output.encoding", "UTF-8");
-        properties.put("resource.loader", "file, classpath, string");
-        properties.put("classpath.resource.loader.class", "org.apache.velocity.runtime.resource.loader.ClasspathResourceLoader");
-        properties.put("string.resource.loader.class", "org.apache.velocity.runtime.resource.loader.StringResourceLoader");
-        properties.put("file.resource.loader.class", "org.apache.velocity.runtime.resource.loader.FileResourceLoader");
-        properties.put("file.resource.loader.path", "/tmp");
-        properties.put("file.resource.loader.cache", false);
-
-        velocityEngineFactoryBean.setVelocityProperties(properties);
-
-        return velocityEngineFactoryBean;
-    }
 }
