+++ settings.gradle
@@ -59,6 +59,7 @@ include ':cas-server-support-mongo'
 include ':cas-server-support-couchbase-core'
 include ':cas-server-support-couchbase-service-registry'
 include ':cas-server-support-couchbase-ticket-registry'
+include ':cas-server-support-saml-core'
 include ':cas-server-support-saml'
 include ':cas-server-support-saml-idp'
 include ':cas-server-support-saml-googleapps'
